rtmlamp: implement reading parameters.
For this purpose, we use a design where one single record actually
triggers the readInt32() method, and that function call leads to reading
all parameters at once from the decoder (cheaper and simpler to do,
since we are reading from memory all at once). It's important to
remember to pass the devinfo struct to the decoder as well.

We also:
- split p_psstatus into its own block, since it's not a per channel
  variable.
- updated the parameter names to the ones used in the DBs and in
  decode-reg.

We should look into using some data structure other than an array of
tuples, which could be indexed by the parameters and allow us to grab
the parameter names directly, without relying on getParamName().
std::unordered_map<int *, const char *> might work.